requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
Django==4.2.7
django-environ==0.11.2
docker==6.1.3
environs==9.5.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.6
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.0
packaging==23.2
pefile==2023.2.7
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
websocket-client==1.6.4
yarl==1.9.4
